Add timer_counters to support |ms type with a sample rate
When feedng samples from OpenStack Swift into Gnocchi, all of
them are rejected with:

ERROR gnocchi.statsd [-] Unable to treat metric
 s02.object-server.REPLICATE.timing:1.90305709839|ms|@0.1:
 Invalid sampling for ms: `0`, should be none

There is no official documentation for StatsD protocol. The
unofficial specification by Benjamin Black says that sample
rate is not allowed for the |ms type. However, StatsD permits
it, and Swift posts events with it. This patch adds a support
for it in Gnocchi too.

The code is stolen from the reference implementation of
StatsD in JavaScript. I have no idea what is going on here.